# api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Shovel Heroes Public API
  version: 0.2.0
  description: |
    REST endpoints replacing previous Base44 SDK usage.
    Comprehensive API documentation including authentication, admin panel, and all CRUD operations.
    Last updated: 2025-10-03 - Added auth, admin, and missing CRUD endpoints.
...
servers:
  - url: https://your.api.server
    description: Production
  - url: http://localhost:8787
    description: Local Dev
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ID:
      type: string
      description: 唯一識別字串 (UUID、ULID 或資料庫主鍵)
      example: "grid_01HZYQ9W123ABCDEF"
...
    User:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/ID'
        name:
          type: string
          description: 使用者顯示名稱
          example: 林志工
        email:
          type: string
          format: email
          description: 使用者電子郵件
          example: volunteer@example.org
      required: [id]
...
    Error:
      type: object
      properties:
        message:
          type: string
          description: 錯誤訊息
          example: 未授權的請求
        code:
          type: string
          description: 內部錯誤代碼 (可選)
          example: AUTH_REQUIRED
      required: [message]
...
  responses:
    Unauthorized:
      description: 未授權 (需要 Bearer Token)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    InternalError:
      description: 伺服器內部錯誤
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
paths:
...
  /users:
    get:
      summary: List users
      description: 取得使用者列表 (需權限)。
      tags: [Users]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/User' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/InternalError' }
  /me:
    get:
      summary: Get current user
      description: 回傳目前登入的使用者資訊。
      tags: [Users]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/InternalError' }
...